requirements.txt
aiohttp==3.8.6
aiosignal==1.3.1
async-timeout==4.0.3
asynctest==0.13.0
attrs==23.2.0
certifi==2024.7.4
charset-normalizer==3.3.2
cycler==0.11.0
datasets==2.13.2
dill==0.3.6
einops==0.6.1
filelock==3.12.2
fonttools==4.38.0
frozenlist==1.3.3
fsspec==2023.1.0
huggingface-hub==0.16.4
idna==3.7
importlib-metadata==6.7.0
kiwisolver==1.4.5
matplotlib==3.5.3
multidict==6.0.5
multiprocess==0.70.14
numpy==1.21.6
nvidia-cublas-cu11==11.10.3.66
nvidia-cuda-nvrtc-cu11==11.7.99
nvidia-cuda-runtime-cu11==11.7.99
nvidia-cudnn-cu11==8.5.0.96
packaging==24.0
pandas==1.3.5
Pillow==9.5.0
pyarrow==12.0.1
pyparsing==3.1.2
python-dateutil==2.9.0.post0
pytz==2024.1
PyYAML==6.0.1
requests==2.31.0
six==1.16.0
torch==1.13.1
torchvision==0.14.1
tqdm==4.66.4
typing_extensions==4.7.1
urllib3==2.0.7
xxhash==3.4.1
yarl==1.9.4
zipp==3.15.0
